pyproject.toml
[project]
name = "nonebot_plugin_weather_lite"
version = "0.0.1"
description = "使用wttr.in的天气查询 ，支持大部分wttr.in的用法。"
authors = [
    {name = "rf_tar_railt", email = "rf_tar_railt@qq.com"},
    {name = "zjkwdy", email = "3377911508@qq.com"},
]
dependencies = [
    "nonebot2>=2.3.0",
    "nonebot-plugin-alconna>=0.45.4"
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}
keywords = ["pip", "nonebot2", "nonebot", "weather", "nonebot_plugin"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
]

[project.urls]
Homepage = "https://github.com/zjkwdy/nonebot_plugin_weather_lite"
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.build]
includes = ["src"]
[tool.pdm.dev-dependencies]
dev = [
    "nonebot-adapter-onebot>=2.4.3",
    "nonebot2[httpx]>=2.3.0",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugin_dirs = ["src"]
builtin_plugins = []